requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
openai>=1.57.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.20
aiosmtplib>=3.0.1
email-validator>=2.2.0
pymongo>=4.10.1
